ID-440 Engooden CreateUserSpec.
